Introduced some enemy animation and random spawning
Tried to do an unlit shader that could be made invisible and faded in-
But I couldn't figure it out.

The idea with these SeraphRails is that they turn in, orient themselves to
the player, and then fire a laser at the player.

Need to come up with a fix to prevent enemies from overlapping.