SERAPH RAILS ARE WORKING SERAPH RAILS ARE WORKING
Will consider a particle system during their wait time
The laser texture is working for now, I'm going to keep it :)

Seraph Rails align themselves and fire, but the code implementing it uses
lots of else ifs, planning to make this a switch case and improve the
scripting later.

Essentially, we have a player and an enemy, so we're well on our way.